% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dat4R}
\alias{example_dat4R}
\title{Example Mating Asymmetry data}
\format{
\subsection{\code{example_dat4R}}{

A dataframe with columns for:
\describe{
\item {type}{Unique identifier for each row}
\item {mt_MS}{mating symmetry factor}
\item {mt_MaS}{mating asymmetry factor}
\item {M}{mother genotype}
\item {F}{father genotype}
\item {C}{child genotype}
\item {E}{environmental effects present (1)/not present (0)}
\item {D}{disease present (1)/not present (0)}
\item {count}{counts for each row of data}
...
}
}
}
\usage{
example_dat4R
}
\description{
Example data formatted to be used in the \code{\link[=MaSlrt]{MaSlrt()}} function.
}
\keyword{datasets}
